Command line option '-v' was added. This option controls verbose progress output.
git-svn-id: https://masu.svn.sourceforge.net/svnroot/masu/trunk@386 5bdb5694-003a-4164-b60f-8fcef0365520

+++ main/src/jp/ac/osaka_u/ist/sel/metricstool/main/MetricsTool.java
@@ -126,7 +126,7 @@ public class MetricsTool {
 
         // ヘルプモードと情報表示モードが同時にオンになっている場合は不正
         if (Settings.isHelpMode() && Settings.isDisplayMode()) {
-            System.err.println("-h and -x can\'t be set at the same time!");
+            err.println("-h and -x can\'t be set at the same time!");
             printUsage();
             System.exit(0);
         }
@@ -164,7 +164,7 @@ public class MetricsTool {
 
         // 対象ファイルのASTから未解決クラス，フィールド，メソッド情報を取得
         {
-            out.println("Parse all target files.");
+            out.println("parsing all target files.");
             final int totalFileNumber = TargetFileManager.getInstance().size();
             int currentFileNumber = 1;
             final StringBuffer fileInformationBuffer = new StringBuffer();
@@ -176,15 +176,17 @@ public class MetricsTool {
                     final FileInfo fileInfo = new FileInfo(name);
                     FileInfoManager.getInstance().add(fileInfo);
 
-                    fileInformationBuffer.delete(0, fileInformationBuffer.length());
-                    fileInformationBuffer.append("parsing ");
-                    fileInformationBuffer.append(name);
-                    fileInformationBuffer.append(" [");
-                    fileInformationBuffer.append(currentFileNumber++);
-                    fileInformationBuffer.append("/");
-                    fileInformationBuffer.append(totalFileNumber);
-                    fileInformationBuffer.append("]");
-                    out.println(fileInformationBuffer.toString());
+                    if (Settings.isVerbose()) {
+                        fileInformationBuffer.delete(0, fileInformationBuffer.length());
+                        fileInformationBuffer.append("parsing ");
+                        fileInformationBuffer.append(name);
+                        fileInformationBuffer.append(" [");
+                        fileInformationBuffer.append(currentFileNumber++);
+                        fileInformationBuffer.append("/");
+                        fileInformationBuffer.append(totalFileNumber);
+                        fileInformationBuffer.append("]");
+                        out.println(fileInformationBuffer.toString());
+                    }
 
                     final Java15Lexer lexer = new Java15Lexer(new FileInputStream(name));
                     final Java15Parser parser = new Java15Parser(lexer);
@@ -212,20 +214,34 @@ public class MetricsTool {
             }
         }
 
-        out.println("Resolve Definitions and Usages.");
-        out.println("STEP1 : resolve class definitions.");
+        out.println("resolving definitions and usages.");
+        if (Settings.isVerbose()) {
+            out.println("STEP1 : resolve class definitions.");
+        }
         registClassInfos();
-        out.println("STEP2 : resolve type parameters of classes.");
+        if (Settings.isVerbose()) {
+            out.println("STEP2 : resolve type parameters of classes.");
+        }
         resolveTypeParameterOfClassInfos();
-        out.println("STEP3 : resolve class inheritances.");
+        if (Settings.isVerbose()) {
+            out.println("STEP3 : resolve class inheritances.");
+        }
         addInheritanceInformationToClassInfos();
-        out.println("STEP4 : resolve field definitions.");
+        if (Settings.isVerbose()) {
+            out.println("STEP4 : resolve field definitions.");
+        }
         registFieldInfos();
-        out.println("STEP5 : resolve method definitions.");
+        if (Settings.isVerbose()) {
+            out.println("STEP5 : resolve method definitions.");
+        }
         registMethodInfos();
-        out.println("STEP6 : resolve method overrides.");
+        if (Settings.isVerbose()) {
+            out.println("STEP6 : resolve method overrides.");
+        }
         addOverrideRelation();
-        out.println("STEP7 : resolve field and method usages.");
+        if (Settings.isVerbose()) {
+            out.println("STEP7 : resolve field and method usages.");
+        }
         addReferenceAssignmentCallRelateion();
 
         // 文法誤りのあるファイル一覧を表示
@@ -240,10 +256,12 @@ public class MetricsTool {
         out.println("finished.");
 
         {
-            /*for (final ClassInfo classInfo : ClassInfoManager.getInstance().getExternalClassInfos()) {
-                out.println(classInfo.getFullQualifiedName(Settings.getLanguage()
-                        .getNamespaceDelimiter()));
-            }*/
+            /*
+             * for (final ClassInfo classInfo :
+             * ClassInfoManager.getInstance().getExternalClassInfos()) {
+             * out.println(classInfo.getFullQualifiedName(Settings.getLanguage()
+             * .getNamespaceDelimiter())); }
+             */
         }
     }
 
@@ -261,7 +279,7 @@ public class MetricsTool {
                 || (!Settings.getFileMetricsFile().equals(Settings.INIT))
                 || (!Settings.getClassMetricsFile().equals(Settings.INIT))
                 || (!Settings.getMethodMetricsFile().equals(Settings.INIT))) {
-            System.err.println("-h can\'t be specified with any other options!");
+            err.println("-h can\'t be specified with any other options!");
             printUsage();
             System.exit(0);
         }
@@ -275,35 +293,35 @@ public class MetricsTool {
     private static void checkDisplayModeParameterValidation() {
         // -d は使えない
         if (!Settings.getTargetDirectory().equals(Settings.INIT)) {
-            System.err.println("-d can\'t be specified in the display mode!");
+            err.println("-d can\'t be specified in the display mode!");
             printUsage();
             System.exit(0);
         }
 
         // -i は使えない
         if (!Settings.getListFile().equals(Settings.INIT)) {
-            System.err.println("-i can't be specified in the display mode!");
+            err.println("-i can't be specified in the display mode!");
             printUsage();
             System.exit(0);
         }
 
         // -F は使えない
         if (!Settings.getFileMetricsFile().equals(Settings.INIT)) {
-            System.err.println("-F can't be specified in the display mode!");
+            err.println("-F can't be specified in the display mode!");
             printUsage();
             System.exit(0);
         }
 
         // -C は使えない
         if (!Settings.getClassMetricsFile().equals(Settings.INIT)) {
-            System.err.println("-C can't be specified in the display mode!");
+            err.println("-C can't be specified in the display mode!");
             printUsage();
             System.exit(0);
         }
 
         // -M は使えない
         if (!Settings.getMethodMetricsFile().equals(Settings.INIT)) {
-            System.err.println("-M can't be specified in the display mode!");
+            err.println("-M can't be specified in the display mode!");
             printUsage();
             System.exit(0);
         }
@@ -320,7 +338,7 @@ public class MetricsTool {
         // -d と -i のどちらも指定されているのは不正
         if (Settings.getTargetDirectory().equals(Settings.INIT)
                 && Settings.getListFile().equals(Settings.INIT)) {
-            System.err.println("-d or -i must be specified in the analysis mode!");
+            err.println("-d or -i must be specified in the analysis mode!");
             printUsage();
             System.exit(0);
         }
@@ -328,14 +346,14 @@ public class MetricsTool {
         // -d と -i の両方が指定されているのは不正
         if (!Settings.getTargetDirectory().equals(Settings.INIT)
                 && !Settings.getListFile().equals(Settings.INIT)) {
-            System.err.println("-d and -i can't be specified at the same time!");
+            err.println("-d and -i can't be specified at the same time!");
             printUsage();
             System.exit(0);
         }
 
         // 言語が指定されなかったのは不正
         if (null == language) {
-            System.err.println("-l must be specified in the analysis mode.");
+            err.println("-l must be specified in the analysis mode.");
             printUsage();
             System.exit(0);
         }
@@ -360,18 +378,18 @@ public class MetricsTool {
 
         // ファイルメトリクスを計測する場合は -F オプションが指定されていなければならない
         if (measureFileMetrics && (Settings.getFileMetricsFile().equals(Settings.INIT))) {
-            System.err.println("-F must be used for specifying a file for file metrics!");
+            err.println("-F must be used for specifying a file for file metrics!");
             System.exit(0);
         }
 
         // クラスメトリクスを計測する場合は -C オプションが指定されていなければならない
         if (measureClassMetrics && (Settings.getClassMetricsFile().equals(Settings.INIT))) {
-            System.err.println("-C must be used for specifying a file for class metrics!");
+            err.println("-C must be used for specifying a file for class metrics!");
             System.exit(0);
         }
         // メソッドメトリクスを計測する場合は -M オプションが指定されていなければならない
         if (measureMethodMetrics && (Settings.getMethodMetricsFile().equals(Settings.INIT))) {
-            System.err.println("-M must be used for specifying a file for method metrics!");
+            err.println("-M must be used for specifying a file for method metrics!");
             System.exit(0);
         }
     }
@@ -400,21 +418,20 @@ public class MetricsTool {
 
         // -l で言語が指定されていない場合は，解析可能言語一覧を表示
         if (null == language) {
-            System.err.println("Available languages;");
+            err.println("Available languages;");
             LANGUAGE[] languages = LANGUAGE.values();
             for (int i = 0; i < languages.length; i++) {
-                System.err.println("\t" + languages[0].getName()
-                        + ": can be specified with term \"" + languages[0].getIdentifierName()
-                        + "\"");
+                err.println("\t" + languages[0].getName() + ": can be specified with term \""
+                        + languages[0].getIdentifierName() + "\"");
             }
 
             // -l で言語が指定されている場合は，そのプログラミング言語で使用可能なメトリクス一覧を表示
         } else {
-            System.err.println("Available metrics for " + language.getName());
+            err.println("Available metrics for " + language.getName());
             for (AbstractPlugin plugin : PluginManager.getInstance().getPlugins()) {
                 PluginInfo pluginInfo = plugin.getPluginInfo();
                 if (pluginInfo.isMeasurable(language)) {
-                    System.err.println("\t" + pluginInfo.getMetricName());
+                    err.println("\t" + pluginInfo.getMetricName());
                 }
             }
             // TODO 利用可能メトリクス一覧を表示
@@ -523,30 +540,30 @@ public class MetricsTool {
      */
     private static void printUsage() {
 
-        System.err.println();
-        System.err.println("Available options:");
-        System.err.println("\t-d: root directory that you are going to analysis.");
-        System.err.println("\t-i: List file including file paths that you are going to analysis.");
-        System.err.println("\t-l: Programming language of the target files.");
-        System.err
-                .println("\t-m: Metrics that you want to get. Metrics names are separated with \',\'.");
-        System.err.println("\t-C: File path that the class type metrics are output");
-        System.err.println("\t-F: File path that the file type metrics are output.");
-        System.err.println("\t-M: File path that the method type metrics are output");
+        err.println();
+        err.println("Available options:");
+        err.println("\t-d: root directory that you are going to analysis.");
+        err.println("\t-i: List file including file paths that you are going to analysis.");
+        err.println("\t-l: Programming language of the target files.");
+        err.println("\t-m: Metrics that you want to get. Metrics names are separated with \',\'.");
+        err.println("\t-v: Output progress verbosely.");
+        err.println("\t-C: File path that the class type metrics are output");
+        err.println("\t-F: File path that the file type metrics are output.");
+        err.println("\t-M: File path that the method type metrics are output");
 
-        System.err.println();
-        System.err.println("Usage:");
-        System.err.println("\t<Help Mode>");
-        System.err.println("\tMetricsTool -h");
-        System.err.println();
-        System.err.println("\t<Display Mode>");
-        System.err.println("\tMetricsTool -x -l");
-        System.err.println("\tMetricsTool -x -l language -m");
-        System.err.println();
-        System.err.println("\t<Analysis Mode>");
-        System.err
+        err.println();
+        err.println("Usage:");
+        err.println("\t<Help Mode>");
+        err.println("\tMetricsTool -h");
+        err.println();
+        err.println("\t<Display Mode>");
+        err.println("\tMetricsTool -x -l");
+        err.println("\tMetricsTool -x -l language -m");
+        err.println();
+        err.println("\t<Analysis Mode>");
+        err
                 .println("\tMetricsTool -d directory -l language -m metrics1,metrics2 -C file1 -F file2 -M file3");
-        System.err
+        err
                 .println("\tMetricsTool -l listFile -l language -m metrics1,metrics2 -C file1 -F file2 -M file3");
     }
 
@@ -554,6 +571,8 @@ public class MetricsTool {
      * 解析対象ファイルを登録
      */
     private static void readTargetFiles() {
+
+        out.println("building target file list.");
 
         // ディレクトリから読み込み
         if (!Settings.getTargetDirectory().equals(Settings.INIT)) {
@@ -693,6 +712,7 @@ public class MetricsTool {
     private static ParameterDef[] parameterDefs = {
             new BoolDef('h', "helpMode", "display usage", true),
             new BoolDef('x', "displayMode", "display available language or metrics", true),
+            new BoolDef('v', "verbose", "output progress verbosely", true),
             new StringDef('d', "targetDirectory", "Target directory"),
             new StringDef('i', "listFile", "List file including paths of target files"),
             new StringDef('l', "language", "Programming language"),
